order-service/requirements.txt
asgiref==3.9.2
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
cryptography==46.0.2
Django==5.2.7
django-cors-headers==4.9.0
djangorestframework==3.16.1
idna==3.10
pika==1.3.2
pycparser==2.23
PyMySQL==1.1.2
python-dotenv==1.1.1
requests==2.32.5
sqlparse==0.5.3
urllib3==2.5.0
